feat: Update project configuration and dependencies
Enable native app deloyment to iPhones in debug mode

Added new framework references and updated the Info.plist to include
necessary keys for improved functionality. This change ensures that
the project is correctly configured to support the latest features
and dependencies, enhancing overall stability and performance.

Additionally, the Podfile.lock was generated to reflect the current
state of dependencies, ensuring that the project can be built
consistently across different environments.